refactor: move fetch_xxx functions from session.py to api.py
